requirements.txt
audioread==3.1.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
deep-audio-features==0.2.18
deprecation==2.1.0
eyeD3==0.9.9
filelock==3.29.0
filetype==1.2.0
fonttools==4.62.1
fsspec==2026.4.0
idna==3.13
imbalanced-learn==0.14.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
MarkupSafe==3.0.3
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.65.1
numpy==2.4.4
packaging==26.2
pandas==2.3.0
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
pyAudioAnalysis==0.3.14
pycparser==3.0
pydub==0.25.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.2
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sklearn-compat==0.1.5
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.9.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3

# Notebook tooling and extras kept from the prior root requirements.txt
# so data_downloads.ipynb / processing.ipynb still install on a fresh clone.
seaborn>=0.13
jupyterlab>=4.2
notebook>=7.2
ipykernel>=6.29
